Add missing :Features parameters. Add shader bitwise operations
Added :Features on GFX.Drawable and GFX.glTF
