+++ src/extra/gfx/shader/math_shards.hpp
@@ -16,6 +16,7 @@
 namespace gfx {
 namespace shader {
 SH_HAS_MEMBER_TEST(validateTypes);
+SH_HAS_MEMBER_TEST(generate);
 SH_HAS_MEMBER_TEST(call);
 SH_HAS_MEMBER_TEST(op);
 SH_HAS_MEMBER_TEST(prefixOp);
@@ -41,25 +42,29 @@ template <typename TShard, typename TOp> struct BinaryOperatorTranslator {
     FieldType typeA = operandA->getType();
     FieldType typeB = operandB->getType();
 
-    FieldType resultType;
-    if constexpr (has_validateTypes<TOp>::value) {
-      resultType = TOp::validateTypes(typeA, typeB);
+    if constexpr (has_generate<TOp>::value) {
+      TOp::generate(std::move(operandA), std::move(operandB), context);
     } else {
-      resultType = typeA;
-      if (typeA != typeB)
-        throw ShaderComposeError(fmt::format("Incompatible operand types left:{}, right:{}", typeA, typeB));
-    }
+      FieldType resultType;
+      if constexpr (has_validateTypes<TOp>::value) {
+        resultType = TOp::validateTypes(typeA, typeB);
+      } else {
+        resultType = typeA;
+        if (typeA != typeB)
+          throw ShaderComposeError(fmt::format("Incompatible operand types left:{}, right:{}", typeA, typeB));
+      }
 
-    if constexpr (has_call<TOp>::value) {
-      // generate `call(A, B)`
-      context.setWGSLTop<WGSLBlock>(
-          resultType, blocks::makeCompoundBlock(TOp::call, "(", operandA->toBlock(), ", ", operandB->toBlock(), ")"));
-    } else if constexpr (has_op<TOp>::value) {
-      // generate `A op B`
-      // store in temp var to avoid precedence issues
-      context.setWGSLTopVar(resultType, blocks::makeCompoundBlock(operandA->toBlock(), TOp::op, operandB->toBlock()));
-    } else {
-      throw std::logic_error("Operator implementation needs to be a call or an operator");
+      if constexpr (has_call<TOp>::value) {
+        // generate `call(A, B)`
+        context.setWGSLTop<WGSLBlock>(
+            resultType, blocks::makeCompoundBlock(TOp::call, "(", operandA->toBlock(), ", ", operandB->toBlock(), ")"));
+      } else if constexpr (has_op<TOp>::value) {
+        // generate `A op B`
+        // store in temp var to avoid precedence issues
+        context.setWGSLTopVar(resultType, blocks::makeCompoundBlock(operandA->toBlock(), TOp::op, operandB->toBlock()));
+      } else {
+        throw std::logic_error("Operator implementation needs to be a call or an operator");
+      }
     }
   }
 };
@@ -131,6 +136,59 @@ struct OperatorDivide {
 
 struct OperatorMod {
   static inline const char *op = "%";
+  static inline FieldType validateTypes(FieldType a, FieldType b) { return validateTypesVectorBroadcast(a, b); }
+};
+
+template <const char *Op>
+inline void generateShift(std::unique_ptr<IWGSLGenerated> &&a, std::unique_ptr<IWGSLGenerated> &&b, TranslationContext &context) {
+  auto typeA = a->getType();
+  auto typeB = b->getType();
+  std::string prefix;
+  std::string suffix;
+
+  // https://www.w3.org/TR/WGSL/#bit-expr
+  // If a is a scalar, b needs to be u32
+  // If a is a vector, b needs to be vec<u32> of that same dimension
+  if (typeB.numComponents != typeA.numComponents) {
+    prefix = fmt::format("{}(u32(", getFieldWGSLTypeName(FieldType(ShaderFieldBaseType::UInt32, typeA.numComponents)));
+    suffix = "))";
+  } else {
+    prefix = "u32(";
+    suffix = ")";
+  }
+
+  context.setWGSLTop<WGSLBlock>(a->getType(),
+                                blocks::makeCompoundBlock("(", a->toBlock(), Op, prefix, b->toBlock(), suffix, ")"));
+}
+
+struct OperatorLShift {
+  static inline void generate(std::unique_ptr<IWGSLGenerated> &&a, std::unique_ptr<IWGSLGenerated> &&b,
+                              TranslationContext &context) {
+    static const char op[] = "<<";
+    generateShift<op>(std::move(a), std::move(b), context);
+  }
+};
+
+struct OperatorRShift {
+  static inline void generate(std::unique_ptr<IWGSLGenerated> &&a, std::unique_ptr<IWGSLGenerated> &&b,
+                              TranslationContext &context) {
+    static const char op[] = ">>";
+    generateShift<op>(std::move(a), std::move(b), context);
+  }
+};
+
+struct OperatorXor {
+  static inline const char *op = "^";
+  static inline FieldType validateTypes(FieldType a, FieldType b) { return validateTypesVectorBroadcast(a, b); }
+};
+
+struct OperatorAnd {
+  static inline const char *op = "&";
+  static inline FieldType validateTypes(FieldType a, FieldType b) { return validateTypesVectorBroadcast(a, b); }
+};
+
+struct OperatorOr {
+  static inline const char *op = "|";
   static inline FieldType validateTypes(FieldType a, FieldType b) { return validateTypesVectorBroadcast(a, b); }
 };
